Interfaz implementada para la descaraga

+++ src/main/java/edu/avanzada/mp3/control/ControladorServidor.java
@@ -24,7 +24,7 @@ public class ControladorServidor {
         return cancionDAO.obtenerCancionesDisponibles();
     }
 
-    public void actualizarEstadoPago(String usuario, double nuevoPago) throws SQLException {
+    public void actualizarEstadoPago(String usuario, int nuevoPago) throws SQLException {
         clienteDAO.actualizarEstadoPago(usuario, nuevoPago);
     }
 }
